Answer the journal's contracts from lineage, identity and the stored row
Nine restored contracts failed on Postgres. Every one was the product, not the
port: the tests assert behaviour the SQLite store had, and each failure is a
place the Postgres implementation answers a narrower question than the one it
was asked.

LINEAGE. contains_message and latest_completed_context_compaction queried one
session's rows. A fork does not own the prefix it inherited -- those rows belong
to the parent -- so a fork was told a message it demonstrably carries is absent,
and that a conversation already compacted has no boundary. The first is worse
than it looks: callers ask it to decide whether a prompt was already admitted,
so a false negative admits it twice. Both now walk the lineage, bounded by the
cut and restricted to fork-inheritable events, and the inherited compaction is
re-projected into the reading session's identity so a caller sees one numbering.

FORKING A ROUTE. fork_before resolved the fork's harness route and compared it
to the parent's, which a fork can never satisfy: resolution asks what a
session's own shape implies, and a fork has a parent, so it is not an empty
root. Forking a native Codex or OpenCode session therefore failed outright,
advising the user to "start a new child session". A fork now COPIES the parent's
decision, as it inherits the transcript that decision belongs to. An undecided
parent still leaves the fork undecided.

CHILD WORKSPACES. register_child_session inserted the owner's workspace with
`on conflict do nothing`, but the row it conflicted with was the child's own
default binding, written moments earlier by create_session. The insert never
applied, and every child sat in a workspace of one while believing it had joined
the team -- which is what broke subagent membership downstream. The update is
now conditional on the binding still being that default, so an inheriting child
is re-homed and a child that deliberately attached elsewhere is not.

VALUES THAT WERE NEVER STORED. attach_workspace and create_session_in_workspace_as
returned the timestamp they bound, but Postgres keeps timestamptz to
microseconds and chrono carries nanoseconds, so the binding handed back could
never equal a later read of the same row. Both now return the stored value.
append_live had the same shape and a worse effect: it stored the coalesced
snapshot and returned the caller's fresh frame, so a client rendering the return
value saw the turn's timer restart on every delta -- the exact effect the
coalescing exists to prevent.

A LISTING THAT OFFERED WHAT IT COULD NOT OPEN. list_sessions rendered a session
whose provider is gone with an empty state. An empty state reads as resumable,
so the picker offered a session that cannot open. Unreadable projections are
skipped again, which still lists every other session -- the concern the previous
behaviour was reaching for.

AND ONE PLAIN BUG. History payload expansion bound the length to
substring(bytea, integer, BIGINT), an overload Postgres does not have, so every
payload-expanding history read failed with "function does not exist".

+++ crates/borg-agent-runtime/src/session_store/postgres/search.rs
@@ -648,7 +648,12 @@ impl PostgresSessionStore {
                     "select payload_kind, byte_len, substring(payload from 1 for $1) as payload \
                      from session_payloads where id = $2",
                 )
-                .bind(i64::try_from(take).unwrap_or(i64::MAX))
+                // `substring(bytea from int for int)` is the only bytea
+                // overload Postgres has: passing a bigint here resolves to no
+                // function at all and fails the whole read. The cap is a byte
+                // budget that never approaches i32::MAX, so the clamp is a
+                // formality rather than a silent truncation.
+                .bind(i32::try_from(take).unwrap_or(i32::MAX))
                 .bind(reference.id)
                 .fetch_optional(self.pool())
                 .await?
